+++ Cargo.toml
@@ -8,7 +8,7 @@ edition = "2021"
 anyhow = "1.0"
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
-# box dyn serialize
+# for box dyn serialize
 erased-serde = "0.3"
 structopt = "0.3"
 # cheap hash for like vec<u8>
