+++ src/expander/bitman.rs
@@ -57,7 +57,7 @@ mod tests {
     use fnv::FnvHashSet;
     use fxhash::FxHashSet;
 
-    use crate::expander::WrappedAHashSet;
+    use crate::expander::set::WrappedAHashSet;
 
     use super::*;
     #[test]
@@ -158,6 +158,36 @@ mod tests {
         assert_eq!(
             BitManipulatorExpander::<WrappedAHashSet<u128>>::expand(parsed_set).len(),
             17
+        );
+    }
+
+    #[test]
+    fn test_1_serialize() {
+        let parsed_set = vec![
+            JsonSet { set: vec![1, 2, 3] },
+            JsonSet { set: vec![4, 5, 6] },
+        ];
+        let expanded_set = BitManipulatorExpander::<FnvHashSet<u128>>::expand(parsed_set);
+        let serialized_set = serde_json::to_string(&expanded_set).unwrap();
+        assert_eq!(
+            serialized_set.len(),
+            35
+        );
+    }
+
+    #[test]
+    fn test_2_serialize() {
+        let parsed_set = vec![
+            JsonSet {
+                set: vec![57, 58, 59, 60],
+            },
+            JsonSet { set: vec![60, 99] },
+        ];
+        let expanded_set = BitManipulatorExpander::<FnvHashSet<u128>>::expand(parsed_set);
+        let serialized_set = serde_json::to_string(&expanded_set).unwrap();
+        assert_eq!(
+            serialized_set.len(),
+            349
         );
     }
 }
