+++ src/expander/bitvec.rs
@@ -2,7 +2,7 @@ use crate::{expander::Expander, JsonSet};
 
 use bitvec::vec::BitVec;
 
-use super::WrappedBitVec;
+use crate::expander::set::WrappedBitVec;
 
 fn convert_itemset(itemset: &[u8], nb_bits: usize) -> WrappedBitVec {
     let mut bv = BitVec::with_capacity(nb_bits);
@@ -70,11 +70,10 @@ mod tests {
 
     use std::collections::HashSet;
 
-    use ahash::AHashSet;
     use fnv::FnvHashSet;
     use fxhash::FxHashSet;
 
-    use crate::expander::WrappedAHashSet;
+    use crate::expander::set::WrappedAHashSet;
 
     use super::*;
     #[test]
@@ -175,6 +174,36 @@ mod tests {
         assert_eq!(
             BitVecExpander::<WrappedAHashSet<WrappedBitVec>>::expand(parsed_set).len(),
             17
+        );
+    }
+
+    #[test]
+    fn test_1_serialize() {
+        let parsed_set = vec![
+            JsonSet { set: vec![1, 2, 3] },
+            JsonSet { set: vec![4, 5, 6] },
+        ];
+        let expanded_set = BitVecExpander::<FnvHashSet<WrappedBitVec>>::expand(parsed_set);
+        let serialized_set = serde_json::to_string(&expanded_set).unwrap();
+        assert_eq!(
+            serialized_set.len(),
+            77
+        );
+    }
+
+    #[test]
+    fn test_2_serialize() {
+        let parsed_set = vec![
+            JsonSet {
+                set: vec![57, 58, 59, 60],
+            },
+            JsonSet { set: vec![60, 99] },
+        ];
+        let expanded_set = BitVecExpander::<FnvHashSet<WrappedBitVec>>::expand(parsed_set);
+        let serialized_set = serde_json::to_string(&expanded_set).unwrap();
+        assert_eq!(
+            serialized_set.len(),
+            140
         );
     }
 }
